+++ CHANGES.rst
@@ -1,14 +1,14 @@
 Changelog
 =========
 
-0.3.5 (????-??-??)
+0.3.5 (2016-01-29)
 ------------------
 
 - added support for `weka.core.BatchPredictor` to class `Classifier` in module `weka.classifiers`
 - upgraded Weka to revision 12311 (post 3.7.13) to avoid performance bottleneck when using setOptions method
 - fixed class `SetupGenerator` from module `weka.core.classes`
 - added `load_any_file` method to the `weka.core.converters` module
-- ...
+- added `save_any_file` method to the `weka.core.converters` module
 
 
 0.3.4 (2016-01-15)